+++ packages/twenty-front/src/modules/favorites/components/PageFavoriteButton.tsx
